requirements.txt
python-dotenv==1.0.0
python-binance==1.0.17
websockets==12.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.26.2
asyncio==3.4.3
aiohttp==3.9.1
psutil==5.9.6